day23: complete.
go solution was mostly a port of a python one i found online as I was struggling to some up with anything. made a few tweaks and added more comments to try to fully understand it

python solution uses a z3 solver. i thought it was interesting enough to add to repo